[SPARK-42870][CONNECT] Move `toCatalystValue` to `connect-common`
### What changes were proposed in this pull request?
Move `toCatalystValue` to connect-common

### Why are the changes needed?
ml support for SCSC also needs this, see https://github.com/apache/spark/pull/40479/files#r1141524125, so we'd better move it to the `connect-common`

### Does this PR introduce _any_ user-facing change?
no, dev-only

### How was this patch tested?
existing UT

Closes #40485 from zhengruifeng/connect_common_proto_value.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
